fix: 深度审查遗留 - 特征真实化(0002迁移)、Celery真实心跳、PII密钥生产拒启、限流XFF、ruff/时区纪律收敛 + 314 tests
- P0-4: Employee 新增20个真实特征字段 + alembic 0002 迁移 + feature_provider 真实字段优先(缺失回退训练分布占位, 无随机)
- P1-6: /health celery 由硬编码改为 Redis 心跳真实探测 + beat 任务
- P0-3: APP_ENV=production 拒绝占位 JWT/PII 密钥
- P0-2: 限流客户端IP解析支持 X-Forwarded-For
- P0-1: warnings router logger 未定义修复
- 工程化: [tool.ruff.lint] 显式规则集 + DTZ011 业务时区(Asia/Shanghai) + 迁移契约测试

+++ backend/app/ml/fusion_engine.py
@@ -93,7 +93,7 @@ class FusionEngine:
         """单员工预测，返回 {risk_score, risk_level, modality_scores}."""
         score_final, modality = self.fuse(structured_features, behavior_features)
         sf = float(score_final[0])
-        risk_score = int(round(sf * 100))
+        risk_score = round(sf * 100)
         return {
             "risk_score": risk_score,
             "risk_level": score_to_level(risk_score),
@@ -108,7 +108,7 @@ class FusionEngine:
         score_final, modality = self.fuse(structured_features, behavior_features)
         results = []
         for i, sf in enumerate(score_final):
-            risk_score = int(round(sf * 100))
+            risk_score = round(sf * 100)
             results.append({
                 "risk_score": risk_score,
                 "risk_level": score_to_level(risk_score),
